requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
dj-database-url==2.2.0
Django==4.2.16
django-allauth==65.0.2
django-crispy-forms==2.3
django-filter==24.3
django-summernote==0.8.20.0
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.1
